resolving issue with race condition
there is a race condition between when the new entry is inserted in line 333 and when it is looked up in lines 340, 345, and 350, as multiple rows can be inserted in between, and cause the subquery to return multiple values causing the second insert query to fail given step5_domain_query would return multiple values. LAST_INSERT_ID should resolve this.